refactor: add unit-managesubscriptions and unit-multicommunity modules
