--- feature-search/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/feature/search/managesubscriptions/ManageSubscriptionsScreen.kt
@@ -1,296 +1,0 @@
-package com.github.diegoberaldin.raccoonforlemmy.feature.search.managesubscriptions
-
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.Icon
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AddCircle
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.ExpandLess
-import androidx.compose.material.pullrefresh.PullRefreshIndicator
-import androidx.compose.material.pullrefresh.pullRefresh
-import androidx.compose.material.pullrefresh.rememberPullRefreshState
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.rememberTopAppBarState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.text.style.TextAlign
-import cafe.adriel.voyager.core.model.rememberScreenModel
-import cafe.adriel.voyager.core.screen.Screen
-import com.github.diegoberaldin.raccoonforlemmy.core.appearance.theme.Spacing
-import com.github.diegoberaldin.raccoonforlemmy.core.architecture.bindToLifecycle
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.components.FloatingActionButtonMenu
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.components.FloatingActionButtonMenuItem
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.detailopener.api.getDetailOpener
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.CommunityItem
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.CommunityItemPlaceholder
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.MultiCommunityItem
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.Option
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.OptionId
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.di.getFabNestedScrollConnection
-import com.github.diegoberaldin.raccoonforlemmy.core.navigation.di.getNavigationCoordinator
-import com.github.diegoberaldin.raccoonforlemmy.core.utils.compose.onClick
-import com.github.diegoberaldin.raccoonforlemmy.core.utils.compose.rememberCallback
-import com.github.diegoberaldin.raccoonforlemmy.core.utils.compose.rememberCallbackArgs
-import com.github.diegoberaldin.raccoonforlemmy.feature.search.di.getManageSubscriptionsViewModel
-import com.github.diegoberaldin.raccoonforlemmy.feature.search.multicommunity.detail.MultiCommunityScreen
-import com.github.diegoberaldin.raccoonforlemmy.feature.search.multicommunity.editor.MultiCommunityEditorScreen
-import com.github.diegoberaldin.raccoonforlemmy.resources.MR
-import dev.icerock.moko.resources.compose.stringResource
-import kotlinx.coroutines.launch
-
-class ManageSubscriptionsScreen : Screen {
-    @OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterialApi::class)
-    @Composable
-    override fun Content() {
-        val model = rememberScreenModel { getManageSubscriptionsViewModel() }
-        model.bindToLifecycle(key)
-        val uiState by model.uiState.collectAsState()
-        val navigatorCoordinator = remember { getNavigationCoordinator() }
-        val topAppBarState = rememberTopAppBarState()
-        val scrollBehavior = TopAppBarDefaults.enterAlwaysScrollBehavior(topAppBarState)
-        val lazyListState = rememberLazyListState()
-        val scope = rememberCoroutineScope()
-        val fabNestedScrollConnection = remember { getFabNestedScrollConnection() }
-        val isFabVisible by fabNestedScrollConnection.isFabVisible.collectAsState()
-        val detailOpener = remember { getDetailOpener() }
-
-        Scaffold(
-            topBar = {
-                TopAppBar(
-                    title = {
-                        Text(
-                            modifier = Modifier.padding(horizontal = Spacing.s),
-                            text = stringResource(MR.strings.navigation_drawer_title_subscriptions),
-                            style = MaterialTheme.typography.titleLarge
-                        )
-                    },
-                    scrollBehavior = scrollBehavior,
-                    navigationIcon = {
-                        Image(
-                            modifier = Modifier.onClick(
-                                onClick = rememberCallback {
-                                    navigatorCoordinator.popScreen()
-                                },
-                            ),
-                            imageVector = Icons.Default.ArrowBack,
-                            contentDescription = null,
-                            colorFilter = ColorFilter.tint(MaterialTheme.colorScheme.onBackground),
-                        )
-                    },
-                )
-            },
-            floatingActionButton = {
-                AnimatedVisibility(
-                    visible = isFabVisible,
-                    enter = slideInVertically(
-                        initialOffsetY = { it * 2 },
-                    ),
-                    exit = slideOutVertically(
-                        targetOffsetY = { it * 2 },
-                    ),
-                ) {
-                    FloatingActionButtonMenu(
-                        items = buildList {
-                            this += FloatingActionButtonMenuItem(
-                                icon = Icons.Default.ExpandLess,
-                                text = stringResource(MR.strings.action_back_to_top),
-                                onSelected = rememberCallback {
-                                    scope.launch {
-                                        lazyListState.scrollToItem(0)
-                                        topAppBarState.heightOffset = 0f
-                                        topAppBarState.contentOffset = 0f
-                                    }
-                                },
-                            )
-                        }
-                    )
-                }
-            },
-        ) { paddingValues ->
-            val pullRefreshState = rememberPullRefreshState(
-                refreshing = uiState.refreshing,
-                onRefresh = rememberCallback(model) {
-                    model.reduce(ManageSubscriptionsMviModel.Intent.Refresh)
-                },
-            )
-            Box(
-                modifier = Modifier
-                    .padding(paddingValues)
-                    .nestedScroll(scrollBehavior.nestedScrollConnection)
-                    .nestedScroll(fabNestedScrollConnection)
-                    .pullRefresh(pullRefreshState),
-            ) {
-                LazyColumn(
-                    modifier = Modifier.fillMaxSize(),
-                    state = lazyListState,
-                    verticalArrangement = Arrangement.spacedBy(Spacing.xxs),
-                ) {
-                    item {
-                        Row(
-                            modifier = Modifier.fillMaxWidth().padding(horizontal = Spacing.s),
-                            verticalAlignment = Alignment.CenterVertically,
-                        ) {
-                            Text(
-                                text = stringResource(MR.strings.manage_subscriptions_header_multicommunities),
-                                style = MaterialTheme.typography.titleLarge,
-                                color = MaterialTheme.colorScheme.onBackground,
-                            )
-                            Spacer(modifier = Modifier.weight(1f))
-                            Icon(
-                                modifier = Modifier.onClick(
-                                    onClick = rememberCallback {
-                                        navigatorCoordinator.pushScreen(
-                                            MultiCommunityEditorScreen()
-                                        )
-                                    },
-                                ),
-                                imageVector = Icons.Default.AddCircle,
-                                contentDescription = null,
-                                tint = MaterialTheme.colorScheme.onBackground,
-                            )
-                        }
-                    }
-                    items(uiState.multiCommunities) { community ->
-                        MultiCommunityItem(
-                            modifier = Modifier.fillMaxWidth()
-                                .background(MaterialTheme.colorScheme.background).onClick(
-                                    onClick = rememberCallback {
-                                        navigatorCoordinator.pushScreen(
-                                            MultiCommunityScreen(community),
-                                        )
-                                    },
-                                ),
-                            community = community,
-                            autoLoadImages = uiState.autoLoadImages,
-                            options = buildList {
-                                this += Option(
-                                    OptionId.Edit,
-                                    stringResource(MR.strings.post_action_edit),
-                                )
-                                this += Option(
-                                    OptionId.Delete,
-                                    stringResource(MR.strings.community_action_unsubscribe),
-                                )
-                            },
-                            onOptionSelected = rememberCallbackArgs(model) { optionId ->
-                                when (optionId) {
-                                    OptionId.Edit -> {
-                                        navigatorCoordinator.pushScreen(
-                                            MultiCommunityEditorScreen(community),
-                                        )
-                                    }
-
-                                    OptionId.Delete -> {
-                                        model.reduce(
-                                            ManageSubscriptionsMviModel.Intent.DeleteMultiCommunity(
-                                                (community.id ?: 0).toInt()
-                                            ),
-                                        )
-                                    }
-
-                                    else -> Unit
-                                }
-                            },
-                        )
-                    }
-                    item {
-                        Row(
-                            modifier = Modifier.fillMaxWidth().padding(horizontal = Spacing.s),
-                            verticalAlignment = Alignment.CenterVertically,
-                        ) {
-                            Text(
-                                text = stringResource(MR.strings.manage_subscriptions_header_subscriptions),
-                                style = MaterialTheme.typography.titleLarge,
-                                color = MaterialTheme.colorScheme.onBackground,
-                            )
-                        }
-                    }
-                    if (uiState.initial) {
-                        items(5) {
-                            CommunityItemPlaceholder()
-                        }
-                    }
-                    items(uiState.communities) { community ->
-                        CommunityItem(
-                            modifier = Modifier.fillMaxWidth()
-                                .background(MaterialTheme.colorScheme.background)
-                                .onClick(
-                                    onClick = rememberCallback {
-                                        detailOpener.openCommunityDetail(community)
-                                    },
-                                ),
-                            community = community,
-                            autoLoadImages = uiState.autoLoadImages,
-                            options = buildList {
-                                this += Option(
-                                    OptionId.Delete,
-                                    stringResource(MR.strings.community_action_unsubscribe),
-                                )
-                            },
-                            onOptionSelected = rememberCallbackArgs(model) { optionId ->
-                                when (optionId) {
-                                    OptionId.Delete -> {
-                                        model.reduce(
-                                            ManageSubscriptionsMviModel.Intent.Unsubscribe(community.id),
-                                        )
-                                    }
-
-                                    else -> Unit
-                                }
-                            }
-                        )
-                    }
-
-                    if (uiState.multiCommunities.isEmpty() && uiState.communities.isEmpty() && !uiState.initial) {
-                        item {
-                            Text(
-                                modifier = Modifier.fillMaxWidth().padding(top = Spacing.xs),
-                                textAlign = TextAlign.Center,
-                                text = stringResource(MR.strings.message_empty_list),
-                                style = MaterialTheme.typography.bodyLarge,
-                                color = MaterialTheme.colorScheme.onBackground,
-                            )
-                        }
-                    }
-                }
-
-                if (!uiState.initial) {
-                    PullRefreshIndicator(
-                        refreshing = uiState.refreshing,
-                        state = pullRefreshState,
-                        modifier = Modifier.align(Alignment.TopCenter),
-                        backgroundColor = MaterialTheme.colorScheme.background,
-                        contentColor = MaterialTheme.colorScheme.onBackground,
-                    )
-                }
-            }
-        }
-    }
-}
