+++ settings.gradle.kts
@@ -62,3 +62,5 @@ include(":unit-userdetail")
 include(":unit-web")
 include(":unit-zoomableimage")
 include(":unit-postlist")
+include(":unit-managesubscriptions")
+include(":unit-multicommunity")
