+++ unit-managesubscriptions/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/managesubscriptions/di/Utils.kt
@@ -1,0 +1,9 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.managesubscriptions.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.managesubscriptions.ManageSubscriptionsMviModel
+import org.koin.java.KoinJavaComponent
+
+actual fun getManageSubscriptionsViewModel(): ManageSubscriptionsMviModel {
+    val res: ManageSubscriptionsMviModel by KoinJavaComponent.inject(ManageSubscriptionsMviModel::class.java)
+    return res
+}
