+++ unit-managesubscriptions/src/iosMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/managesubscriptions/di/Utils.kt
@@ -1,0 +1,12 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.managesubscriptions.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.managesubscriptions.ManageSubscriptionsMviModel
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
+
+actual fun getManageSubscriptionsViewModel(): ManageSubscriptionsMviModel =
+    UnitManageSubscriptionsDiHelper.manageSuscriptionsViewModel
+
+object UnitManageSubscriptionsDiHelper : KoinComponent {
+    val manageSuscriptionsViewModel: ManageSubscriptionsMviModel by inject()
+}
