+++ unit-multicommunity/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/multicommunity/di/MultiCommunityModule.kt
@@ -1,0 +1,48 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.multicommunity.di
+
+import com.github.diegoberaldin.raccoonforlemmy.core.architecture.DefaultMviModel
+import com.github.diegoberaldin.raccoonforlemmy.unit.multicommunity.detail.MultiCommunityMviModel
+import com.github.diegoberaldin.raccoonforlemmy.unit.multicommunity.detail.MultiCommunityViewModel
+import com.github.diegoberaldin.raccoonforlemmy.unit.multicommunity.editor.MultiCommunityEditorMviModel
+import com.github.diegoberaldin.raccoonforlemmy.unit.multicommunity.editor.MultiCommunityEditorViewModel
+import com.github.diegoberaldin.raccoonforlemmy.unit.multicommunity.utils.DefaultMultiCommunityPaginator
+import com.github.diegoberaldin.raccoonforlemmy.unit.multicommunity.utils.MultiCommunityPaginator
+import org.koin.dsl.module
+
+val multiCommunityModule = module {
+    factory<MultiCommunityMviModel> { params ->
+        MultiCommunityViewModel(
+            mvi = DefaultMviModel(MultiCommunityMviModel.UiState()),
+            community = params[0],
+            postRepository = get(),
+            identityRepository = get(),
+            apiConfigurationRepository = get(),
+            siteRepository = get(),
+            themeRepository = get(),
+            shareHelper = get(),
+            settingsRepository = get(),
+            notificationCenter = get(),
+            hapticFeedback = get(),
+            paginator = get(),
+            imagePreloadManager = get(),
+            getSortTypesUseCase = get(),
+        )
+    }
+    factory<MultiCommunityPaginator> {
+        DefaultMultiCommunityPaginator(
+            postRepository = get(),
+        )
+    }
+    factory<MultiCommunityEditorMviModel> { params ->
+        MultiCommunityEditorViewModel(
+            mvi = DefaultMviModel(MultiCommunityEditorMviModel.UiState()),
+            editedCommunity = params[0],
+            identityRepository = get(),
+            communityRepository = get(),
+            accountRepository = get(),
+            multiCommunityRepository = get(),
+            notificationCenter = get(),
+            settingsRepository = get(),
+        )
+    }
+}
